# docs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "GAMA STORE REST API",
    "description": "API Documentation for GAMA STORE",
    "contact": {
      "name": "@hectormr206",
      "url": "https://hectormr.com",
      "mail": "contact@hectormr.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3003/"
    }
  ],
  "paths": {
...
    "/api/v1/products/measures/": {
      "get": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Get all brands",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Create new brand",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductMeasure"
              },
              "examples": {
                "ProductMeasure": {
                  "$ref": "#/components/examples/ProductMeasure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/measures/{id}": {
      "get": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Get brand by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Update brand by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductMeasure"
              },
              "examples": {
                "ProductMeasure": {
                  "$ref": "#/components/examples/ProductMeasure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/categories/": {
      "get": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Get all brands",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Create new brand",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCategory"
              },
              "examples": {
                "ProductCategory": {
                  "$ref": "#/components/examples/ProductCategory"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/categories/{id}": {
      "get": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Get brand by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Update brand by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCategory"
              },
              "examples": {
                "ProductCategory": {
                  "$ref": "#/components/examples/ProductCategory"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/brands/": {
      "get": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Get all brands",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Create new brand",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBrand"
              },
              "examples": {
                "ProductBrand": {
                  "$ref": "#/components/examples/ProductBrand"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/brands/{id}": {
      "get": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Get brand by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Products (Brands)"
        ],
        "summary": "Update brand by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBrand"
              },
              "examples": {
                "ProductBrand": {
                  "$ref": "#/components/examples/ProductBrand"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create new product",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "Product": {
                  "$ref": "#/components/examples/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "Update product by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "Product": {
                  "$ref": "#/components/examples/Product"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "examples": {
...
      "ProductMeasure": {
        "value": {
          "name": "Legal"
        }
      },
      "ProductCategory": {
        "value": {
          "name": "Café"
        }
      },
      "ProductBrand": {
        "value": {
          "name": "Pieza"
        }
      },
      "Product": {
        "value": {
          "name": "Legal sobre 30 gr",
          "universalCode": "7501206613733",
          "productBrandId": 1,
          "productCategoryId": 2,
          "productMeasureId": 1
        }
      }
    },
    "schemas": {
...
      "ProductMeasure": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Legal"
          }
        }
      },
      "ProductCategory": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Café"
          }
        }
      },
      "ProductBrand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Pieza"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Legal sobre 30 gr"
          },
          "universalCode": {
            "type": "string",
            "example": "7501206613733"
          },
          "productBrandId": {
            "type": "number",
            "example": 1
          },
          "productCategoryId": {
            "type": "number",
            "example": 2
          },
          "productMeasureId": {
            "type": "number",
            "example": 1
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}